[REF] web: convert public component service tests to hoot
task-3705027
